"Enhanced clarity in endpoint print message and added tunneling command"
This commit modified the print statement for the endpoint in order to increase readability and clarity. The name of the endpoint is now highlighted separately to make it more distinguishable in the output.

Additionally, a new command has been registered, 'tunnel', in discoveryrotation.py. This is a PAMTunnelCommand which will help manage tunnels for the application. However, some features of this command are still under development as indicated by the commented lines. This new command broadens the functionality of our application, specifically providing more control over the tunneling feature.