Allow "hidden" output
A new clipboard-copy output option "stdouthidden" which prints
the password "red on red" to the terminal.
Additonally a "sh" - "show hidden" alias.
One backdraw if the terminal supports highlight colors the password will
be unveiled on mouse selection.

Usecase: Sometime it is handy to have the password in mouse buffer in
X-Windows/Wayland systems.

Idea shameless stolen from kpcli.