DR-667 Fix problem with item count and now automatically remove finished jobs.
The old item count was broken, a quick last momemnt addtion. Fix it
so current number is correct.

When the customer finishes processing a discovery job, it will automatically
remove the job. If the discovery results in nothing to discover, it will automatically
remove the job.

Make sure all wheel files are update to date.