DR-854, DR-866 local pamUser assignment and setting resourceUid to blank.
DR-854 Don't allow local pamUser record reuse.

During processing discovery, if existing local pamUser record
is assigned to the resource don't let the same local pamUser record
be used for another resource.

The list of pamUser records to be converted are passed to
'find user' method. When we filter out users, use it to
filter out/notifiy it's taken when list users.

DR-866 Don't assigned resourceUid is IAM User.

If the parent of the user is a provider, make sure the
resourceUid for set rotation is None.